+++ src/routes/applicationNew.js
@@ -3,6 +3,7 @@ import { Grid, Card, CardContent, Typography, TextField, RadioGroup, FormControl
 import { Portal } from '@mui/base';
 
 import { customAxios as axios, getAuthToken } from '../functions';
+import { useTheme } from '@emotion/react';
 
 var vars = require("../variables");
 
@@ -13,11 +14,13 @@ var vars = require("../variables");
 //     },
 //     {
 //         "type": "text",
-//         "label": "Name"
+//         "label": "Name",
+//         "min_length": 150
 //     },
 //     {
 //         "type": "textarea",
-//         "label": "Bio"
+//         "label": "Bio",
+//         "min_length": 150
 //     },
 //     {
 //         "type": "number",
@@ -37,10 +40,6 @@ var vars = require("../variables");
 //         "choices": ["English", "Spanish", "French"]
 //     },
 //     {
-//         "type": "position",
-//         "label": "What position are you applying for?" // this syncs to /applications/positions
-//     },
-//     {
 //         "type": "radio",
 //         "label": "Gender",
 //         "choices": ["Male", "Female", "Other"]
@@ -52,13 +51,24 @@ var vars = require("../variables");
 //     }
 // ];
 
-const CustomForm = ({ config, formData, setFormData }) => {
+const CustomForm = ({ theme, config, formData, setFormData, setSubmitDisabled }) => {
     if (config === undefined) return <Typography>Please select the type of application in the dropdown on the top-right corner</Typography>;
 
+    let fieldReq = {};
     let defaultResp = {};
     for (let i = 0; i < config.length; i++) {
         if (config[i].type !== "info") {
             defaultResp[config[i].label] = "";
+            if (["text", "textarea"].includes(config[i].type)) {
+                fieldReq[config[i].label] = { ...fieldReq[config[i].label], min_length: config[i].min_length };
+            }
+            if (["text", "textarea", "date", "datetime", "number", "dropdown", "radio", "checkbox"].includes(config[i].type)) {
+                fieldReq[config[i].label] = { ...fieldReq[config[i].label], must_input: config[i].must_input };
+            }
+            if (["number"].includes(config[i].type)) {
+                fieldReq[config[i].label] = { ...fieldReq[config[i].label], min_value: config[i].min_value, max_value: config[i].max_value };
+            }
+
             if (config[i].type === "date") {
                 defaultResp[config[i].label] = new Date(new Date().getTime() - new Date().getTimezoneOffset() * 60000).toISOString().slice(0, 10);
             }
@@ -83,10 +93,39 @@ const CustomForm = ({ config, formData, setFormData }) => {
 
     const handleChange = (e) => {
         const { name, value } = e.target;
-        setFormData(data => ({
-            ...data,
-            [name]: value
-        }));
+        setFormData(data => ({ ...data, [name]: value }));
+        const newFormData = { ...formData, [name]: value };
+        const formKeys = Object.keys(newFormData);
+        let allOk = true;
+        for (let i = 0; i < formKeys.length; i++) {
+            if (fieldReq[formKeys[i]] !== undefined) {
+                if (fieldReq[formKeys[i]].min_length !== undefined) {
+                    if (newFormData[formKeys[i]].length < fieldReq[formKeys[i]].min_length) {
+                        setSubmitDisabled(true);
+                        allOk = false;
+                    }
+                }
+                if (fieldReq[formKeys[i]].must_input === true) {
+                    if (newFormData[formKeys[i]] === undefined || newFormData[formKeys[i]].replaceAll(" ", "") === "") {
+                        setSubmitDisabled(true);
+                        allOk = false;
+                    }
+                }
+                if (fieldReq[formKeys[i]].min_value !== undefined) {
+                    if (newFormData[formKeys[i]] < fieldReq[formKeys[i]].min_value) {
+                        setSubmitDisabled(true);
+                        allOk = false;
+                    }
+                }
+                if (fieldReq[formKeys[i]].max_value !== undefined) {
+                    if (newFormData[formKeys[i]] > fieldReq[formKeys[i]].max_value) {
+                        setSubmitDisabled(true);
+                        allOk = false;
+                    }
+                }
+            }
+        }
+        if (allOk) setSubmitDisabled(false);
     };
 
     const handleCheckboxChange = (choice, fieldLabel) => {
@@ -140,7 +179,9 @@ const CustomForm = ({ config, formData, setFormData }) => {
                                     value={formData[field.label]}
                                     onChange={handleChange}
                                     placeholder={field.placeholder}
-                                    sx={{ width: "100%" }}
+                                    sx={{ width: "100%", '& .MuiFormHelperText-root': { color: theme.palette.error.main } }}
+                                    error={field.min_length !== undefined && formData[field.label] !== "" && formData[field.label].length <= field.min_length}
+                                    helperText={field.min_length !== undefined && formData[field.label] !== "" && formData[field.label].length <= field.min_length ? "Input at least 150 characters" : ""}
                                 />
                             );
                             break;
@@ -156,7 +197,9 @@ const CustomForm = ({ config, formData, setFormData }) => {
                                     onChange={handleChange}
                                     rows={field.rows}
                                     placeholder={field.placeholder}
-                                    sx={{ width: "100%" }}
+                                    sx={{ width: "100%", '& .MuiFormHelperText-root': { color: theme.palette.error.main } }}
+                                    error={field.min_length !== undefined && formData[field.label] !== "" && formData[field.label].length <= field.min_length}
+                                    helperText={field.min_length !== undefined && formData[field.label] !== "" && formData[field.label].length <= field.min_length ? "Input at least 150 characters" : ""}
                                 />
                             );
                             break;
@@ -168,9 +211,11 @@ const CustomForm = ({ config, formData, setFormData }) => {
                                     name={field.label}
                                     label={field.label}
                                     value={formData[field.label]}
-                                    onChange={handleChange}
-                                    type="number"
-                                    sx={{ width: "100%" }}
+                                    onChange={(e) => { if (!isNaN(e.target.value)) handleChange(e); }}
+                                    type="text"
+                                    sx={{ width: "100%", '& .MuiFormHelperText-root': { color: theme.palette.error.main } }}
+                                    error={field.min_value !== undefined && formData[field.label] < field.min_value || field.max_value !== undefined && formData[field.label] > field.max_value}
+                                    helperText={field.min_value !== undefined && formData[field.label] < field.min_value ? `Minimum value ${field.min_value}` : (field.max_value !== undefined && formData[field.label] < field.max_value ? `Maximum value ${field.max_value}` : "")}
                                 />
                             );
                             break;
@@ -292,10 +337,146 @@ const CustomForm = ({ config, formData, setFormData }) => {
 };
 
 const NewApplication = () => {
+    // FOR LOCAL TESTING PURPOSE
+    if (window.location.pathname === "localhost") {
+        applicationTypes = [
+            {
+                "id": 1,
+                "name": "Driver",
+                "form": [
+                    {
+                        "type": "date",
+                        "label": "What is your birthday?",
+                        "must_input": true
+                    },
+                    {
+                        "type": "textarea",
+                        "label": "How did you find us?",
+                        "rows": "3",
+                        "placeholder": "Enter a short answer",
+                        "min_length": 150
+                    },
+                    {
+                        "type": "radio",
+                        "label": "Are you currently in another VTC?",
+                        "choices": [
+                            "Yes",
+                            "No"
+                        ],
+                        "must_input": true
+                    },
+                    {
+                        "type": "textarea",
+                        "label": "What are your interests?",
+                        "rows": "5",
+                        "placeholder": "Tell us a little bit about yourself",
+                        "min_length": 150
+                    },
+                    {
+                        "type": "textarea",
+                        "label": "Why do you want to be a part of our VTC?",
+                        "rows": "5",
+                        "placeholder": "Why would you like to join us? This doesn't need to be complicated.",
+                        "min_length": 150
+                    },
+                    {
+                        "type": "checkbox",
+                        "label": "By joining the VTC, you agree to follow both discord and VTC rules at all times? Do you agree to our terms?",
+                        "must_input": true
+                    }
+                ]
+            },
+            {
+                "id": 2,
+                "name": "Staff",
+                "form": [
+                    {
+                        "type": "date",
+                        "label": "What is your birthdate?"
+                    },
+                    {
+                        "type": "textarea",
+                        "label": "What country do you live in? Also include your Time Zone",
+                        "placeholder": "US, Canada, UK, etc",
+                        "rows": "3",
+                        "min_length": 150
+                    },
+                    {
+                        "type": "dropdown",
+                        "label": "Which position are your applying for?",
+                        "choices": [
+                            "Events Team"
+                        ],
+                        "must_input": true
+                    },
+                    {
+                        "type": "textarea",
+                        "label": "Please provide a summary about yourself.",
+                        "placeholder": "You may include hobbies, work positions, or any unique facts about yourself!",
+                        "rows": "5",
+                        "min_length": 150
+                    },
+                    {
+                        "type": "textarea",
+                        "label": "Why are you interested in joining the position you are applying for? What do you want to achieve?",
+                        "placeholder": "Explain why does that position interest you and what do you want to achieve.",
+                        "rows": "5",
+                        "min_length": 150
+                    },
+                    {
+                        "type": "textarea",
+                        "label": "Do you have a lot of time to dedicate to this position?",
+                        "placeholder": "Explain your time availability.",
+                        "rows": "3",
+                        "min_length": 150
+                    },
+                    {
+                        "type": "checkbox",
+                        "label": "By joining the VTC, you agree to follow both discord and VTC rules at all times? Do you agree to our terms?",
+                        "must_input": true
+                    }
+                ]
+            },
+            {
+                "id": 3,
+                "name": "LOA",
+                "form": [
+                    {
+                        "type": "date",
+                        "label": "Start Date"
+                    },
+                    {
+                        "type": "date",
+                        "label": "End Date"
+                    },
+                    {
+                        "type": "textarea",
+                        "label": "Reason for LOA",
+                        "rows": "3",
+                        "min_length": 150
+                    },
+                    {
+                        "type": "checkbox",
+                        "label": "I will leave the staff position"
+                    },
+                    {
+                        "type": "checkbox",
+                        "label": "I will leave the VTC"
+                    }
+                ]
+            },
+            {
+                "id": 4,
+                "name": "Division"
+            }
+        ];
+    }
+
+    const theme = useTheme();
     const [selectedType, setSelectedType] = useState(null);
     const listTypes = Object.values(vars.applicationTypes);
     const [enableNotifications, setEnableNotifications] = useState(true);
-    const [submitLoading, setSubmitLoading] = useState(false);
+    const [submitDisabled, setSubmitDisabled] = useState(true);
 
     const [formData, setFormData] = useState(null);
 
@@ -306,7 +487,7 @@ const NewApplication = () => {
     }, []);
 
     const handleSubmit = useCallback(async () => {
-        setSubmitLoading(true);
+        setSubmitDisabled(true);
 
         let modFormData = formData;
         let keys = Object.keys(modFormData);
@@ -331,7 +512,7 @@ const NewApplication = () => {
             setSnackbarSeverity("error");
         }
 
-        setSubmitLoading(false);
+        setSubmitDisabled(false);
     }, [enableNotifications, formData, selectedType]);
 
     return <Card sx={{ padding: "20px" }}>
@@ -353,7 +534,7 @@ const NewApplication = () => {
             </TextField>
         </div>
         <CardContent>
-            <CustomForm config={selectedType !== null ? vars.applicationTypes[selectedType].form : undefined} formData={formData} setFormData={setFormData} />
+            <CustomForm theme={theme} config={selectedType !== null ? vars.applicationTypes[selectedType].form : undefined} formData={formData} setFormData={setFormData} setSubmitDisabled={setSubmitDisabled} />
             {((selectedType !== null ? vars.applicationTypes[selectedType].form : undefined) !== undefined) &&
                 <Box sx={{ display: 'grid', justifyItems: 'end' }}>
                     <div style={{ display: 'flex', alignItems: 'center' }}>
@@ -370,7 +551,7 @@ const NewApplication = () => {
                                 label="Enable notifications?"
                             />
                         </FormControl>
-                        <Button onClick={handleSubmit} variant="contained" color="info" disabled={submitLoading}>
+                        <Button onClick={handleSubmit} variant="contained" color="info" disabled={submitDisabled}>
                             Submit
                         </Button>
                     </div>